third_party/blink/web_tests/platform/mac/external/wpt/html/rendering/widgets/baseline-alignment-and-overflow.tentative-expected.txt

This is a testharness.js-based test.
Found 28 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <input type="date" value="2020-01-01" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="date" value="2020-01-01" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="month" value="2020-01" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="month" value="2020-01" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="week" value="2020-W01" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="week" value="2020-W01" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="time" value="00:00" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="time" value="00:00" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="datetime-local" value="2020-01-01T00:00" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="datetime-local" value="2020-01-01T00:00" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="range" style="overflow: visible; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: hidden; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: visible; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: hidden; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="range" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 70 but got 60
[FAIL] <input type="color" value="#000000" style="overflow: visible; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: hidden; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 70
[FAIL] <input type="color" value="#000000" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 48 but got 70
[FAIL] <input type="color" value="#000000" style="overflow: visible; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 45
[FAIL] <input type="color" value="#000000" style="overflow: hidden; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 70
[FAIL] <input type="color" value="#000000" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 48 but got 70
[FAIL] <input type="submit" value="x" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="submit" value="x" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="reset" value="x" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="reset" value="x" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="button" value="x" style="overflow: scroll; appearance: auto;">
  assert_equals: <span>.offsetTop expected 34 but got 26
[FAIL] <input type="button" value="x" style="overflow: scroll; appearance: none;">
  assert_equals: <span>.offsetTop expected 34 but got 26
Harness: the test ran to completion.
